pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Patrick Arminio", email = "patrick.arminio@gmail.com"},
]
dependencies = [
    "strawberry-graphql>=0.127.3",
    "fastapi>=0.81.0",
    "uvicorn>=0.18.3",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.6.0",
    "pytest>=7.1.2",
    "requests>=2.28.1",
]
